+++ package/tests/api.test.ts
@@ -1,24 +1,47 @@
-import { getDataFromServer, getHistoricals, getStock, getStocks } from '../src/api';
+import { getDataFromServer, getHistoricals, getStock, getStocks, getTicker, getTickers } from '../src';
 
-// getHistoricals('bsp')
-// getStock('bsp')
-// getDataFromServer('BSP').then(res => {
-//     console.log(res)
-// })
-
-// describe('retrieving data from www.pngx.com.pg', () => {
-//     test('should return full data in json', async () => {
-//         const result = await getDataFromServer('BSP');
-//         expect(result).toBe('expected value');
-//     })
-// })
-
-describe('retrieving data from localhost', () => {
+describe('fetches stocks for the current day', () => {
     test('should return full data in json', async () => {
         const result = await getStocks();
         expect(result).toBe('expected value');
     })
-})
+});
+
+describe('fetches stock for BSP', () => {
+    test('should return full data in json', async () => {
+        const result = await getStock("BSP");
+        expect(result).toBe('expected value');
+    })
+});
+
+describe('fetches ticker for BSP', () => {
+    test('should return full data in json', async () => {
+        const result = await getTickers();
+        expect(result).toBe('expected value');
+    })
+});
+
+describe('fetches ticker for BSP', () => {
+    test('should return full data in json', async () => {
+        const result = await getTicker("BSP");
+        expect(result).toBe('expected value');
+    })
+});
+
+describe('fetches historical data for BSP', () => {
+    test('should return full data in json', async () => {
+        const result = await getHistoricals("BSP");
+        expect(result).toBe('expected value');
+    })
+});
+
+describe('fetches data from www.pngx.com.pg', () => {
+    test('should return full data in json', async () => {
+        const result = await getDataFromServer('BSP');
+        expect(result).toBe('expected value');
+    })
+});
+
 
 // test('fetches data from an API', async () => {
 //     // mock the fetch function to return a fake response