+++ routes/index.js
@@ -8,6 +8,23 @@ router.get("/", (req, res) => {
   });
 });
 
+// Health check endpoint
+router.get('/health', async (_req, res, _next) => {
+	// optional: add further things to check (e.g. connecting to dababase)
+	const healthcheck = {
+		uptime: process.uptime(),
+		message: 'OK',
+		timestamp: Date.now()
+	};
+
+	try {
+		res.json(healthcheck);
+	} catch (e) {
+		healthcheck.message = e;
+		res.status(503).send();
+	}
+});
+
 /**
  * /api/v1
  */
@@ -17,6 +34,8 @@ router.use("/v1", require("./v1"));
  * /api/v2
  */
 router.use("/v2", require("./v2"));
+
+// router.use('/webhook', require('./webhook'));
 
 router.all("/*", (req, res) => {
   res.status(404).json({